+++ docs/conf.py
@@ -13,7 +13,7 @@ sys.path.insert(0, os.path.abspath(".."))
 project = "FDFI"
 copyright = "2025, FDFI Team"
 author = "FDFI Team"
-release = "0.0.6"
+release = "0.0.7"
 
 # -- General configuration ---------------------------------------------------
 extensions = [